python/pyproject.toml
[tool.poetry]
name = "onnx_inference_using_opencv"
version = "0.1.0"
description = ""
authors = ["chatflip <oo.chat.flip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7,<3.11"
torch = "^1.9.1"
torchvision = "^0.10.1"
opencv-python = "^4.5.4"
onnx = "^1.10.2"
onnx-simplifier = "^0.3.6"
onnxruntime = "^1.9.0"

[tool.poetry.dev-dependencies]
pysen = "^0.10"
black = "=21.10b0"
flake8 = "=4.0.1"
isort = "=5.10.1"
mypy = "=0.910"
autopep8 = "*"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "very_strict"
line_length = 88
py_version = "py37"

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
